refactor: remove createAIGatewayFromEnv from shared, add per-app route adapters
- Delete createAIGatewayFromEnv() from apps/shared/src/lib/ai-gateway.ts
  (shared code should never assume which app's env it runs in)
- Rewrite frontend chat/health routes to build AIGateway directly from
  process.env, using app's own createClient() for Supabase session
  extraction (fixes cookie refresh bug)
- Add chat and health route adapters for admin-portal (reuses
  supabase-server.ts) and mobile-pwa (reuses supabase/server.ts)
- All three apps verified: TypeScript builds clean, routes registered
