+++ apps/admin-portal/src/app/api/v1/ai/health/route.ts
