+++ apps/mobile-pwa/src/app/api/v1/ai/health/route.ts
